Add demo lambda runtime
My team maintains a project that simulates a local lambda Invoke
endpoint. I've added this in along with some docker files to make it
possible to run a local system that is compatible with the AWS CLI for
invoking lambda functions. This is not a request to add this to a final
version of the project. This is only for demo purposes.